requirements.txt
groq
python-telegram-bot
apscheduler
python-dotenv
psycopg2-binary
openpyxl
matplotlib